ai-service/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# AI/ML
openai==1.12.0
langchain==0.1.6
langchain-openai==0.0.5
tiktoken==0.5.2
numpy==1.26.3

# Speech Processing
openai-whisper==20231117
pydub==0.25.1
speechrecognition==3.10.1
gTTS==2.5.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Caching
redis==5.0.1

# Logging
loguru==0.7.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0

